Break up communication page into components
